requirements.txt
aiohttp>=3.8.0
aiohttp-socks==0.6.0
async-timeout>=4.0.0
attrs==20.3.0
chardet==4.0.0
hexdump==3.3
idna==3.1
multidict==5.1.0
python-socks==1.2.4
typing-extensions>=4.0.0
yarl==1.6.3
python-dotenv
fastapi
uvicorn
redis>=4.2.0